Moved AnimatedWidget to pyqt wrappers
