+++ shortcut_composer/api_krita/pyqt/__init__.py
@@ -4,6 +4,7 @@
 """Wrappers and utilities based on PyQt5 objects."""
 
 from .pixmap_transform import PixmapTransform
+from .animated_widget import AnimatedWidget
 from .colorizer import Colorizer
 from .painter import Painter
 from .timer import Timer
@@ -11,6 +12,7 @@ from .text import Text
 
 __all__ = [
     "PixmapTransform",
+    "AnimatedWidget",
     "Colorizer",
     "Painter",
     "Timer",
