+++ shortcut_composer/templates/pie_menu_utils/pie_widget.py
@@ -4,42 +4,12 @@
 from typing import List
 
 from PyQt5.QtCore import Qt, QPoint
-from PyQt5.QtWidgets import QWidget
 from PyQt5.QtGui import QColor, QPaintEvent
 
-from api_krita.pyqt import Painter, Timer
-from composer_utils import Config
+from api_krita.pyqt import AnimatedWidget, Painter
 from .pie_style import PieStyle
 from .label import LabelPainter
 from .label_holder import LabelHolder
-
-
-class AnimatedWidget(QWidget):
-    """Adds the fade-in animation when the widget is shown (60 FPS)."""
-
-    def __init__(self, parent) -> None:
-        super().__init__(parent)
-        self._animation_interval = self._read_animation_interval()
-        self._animation_timer = Timer(self._increase_opacity, 17)
-
-    def _increase_opacity(self):
-        """Add interval to current opacity, stop the timer when full."""
-        current_opacity = self.windowOpacity()
-        self.setWindowOpacity(current_opacity+self._animation_interval)
-        if current_opacity >= 1:
-            self._animation_timer.stop()
-
-    def show(self):
-        """Decrease opacity to 0, and start a timer which animates it."""
-        self.setWindowOpacity(0)
-        self._animation_timer.start()
-        super().show()
-
-    def _read_animation_interval(self):
-        """Return how much opacity (0-1) should be increased on each frame."""
-        if time := Config.PIE_ANIMATION_TIME.read():
-            return 0.0167/time
-        return 1
 
 
 class PieWidget(AnimatedWidget):
